requirements.txt
autogen-core
autogen-magentic-one
autogen-ext
python-dotenv
playwright
pillow
requests
stripe
google-api-python-client